+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * add `afragen/add-plugin-dependency-api` as composer requirement
 * update test plugins removing `hello-dolly` and adding `git-updater` as non dot org example
+* check empty plugin response for error
 
 #### 1.4.1 / 2022-08-18
 * oops, fixed typo in one of the testing plugins